Try making an abstract task
